Add files via upload
This is a tricky one. I want to try and answer your question using three types of historic data that could indicate where America’s worst drivers are: The number of car crashes in each state (especially those where the driver was negligent in some way), how much insurance companies pay out, and how much insurance companies charge drivers. All three measures vary a lot across the country and no state is consistently at the bottom, but drivers in Texas don’t do very well on any of them. By contrast, drivers in Iowa, Indiana and Vermont are pretty good across the board.
#https://www.martinandlerda.com/steps-to-take-if-youve-been-in-a-car-accident/

Recent Stories from FiveThirtyEight
What’s Behind Biden’s Record-Low Approval Rating?
First, collisions. There were 5.6 million motor vehicle crashes in the United States in 2012, of which 4 million involved solely property damage, 1.6 million involved a personal injury, and 30,800 resulted in a fatality, according to the National Highway Traffic Safety Administration (NHTSA). Texas accounted for 3,021 of those fatal crashes, more than any other state, while Washington, D.C., had 14, fewer than anywhere else.

But a state with more fatal car crashes is not necessarily a state with more bad drivers — it could simply be a state that has more drivers or a state with worse driving conditions. So we need to factor in the total miles traveled in each state and focus on the characteristics of the drivers who were involved in fatal collisions.

North Dakota has the highest number of such drivers for every billion miles traveled. Over the course of the 9.1 billion miles traveled in the state in 2011 (the latest data available), 218 drivers were involved in 147 crashes. That produces a figure of 23.8 drivers involved in fatal collisions for every billion miles traveled — far higher than the national average of 15.5.

You’re interested in the behavior of those individuals. We know, for example, whether a driver was distracted at the time of a fatal accident. Of the 45,670 drivers involved in fatal road accidents nationwide in 2012, 3,758 were recorded as being distracted at the time (although it’s worth bearing in mind that in 8,991 cases, it was either not known or not recorded whether the driver was distracted). There’s detail on what those distractions were: 397 of those drivers were distracted by their cellphones, 39 were eating or drinking, and 17 drivers were simply “lost in thought/day dreaming.”

In both Ohio and North Dakota, just 1 percent of drivers involved in fatal accidents were recorded as distracted, compared to 10 percent nationally. But those percentages need to be treated with plenty of caution — it might sound pretty impressive that none of the drivers involved in fatal accidents in D.C. were recorded as being distracted at the time, but that’s based on only 10 drivers for which we have information.